backend/requirements.txt
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pypdf>=3.0.0
python-docx>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pytz>=2024.1
sqlalchemy>=2.0.0
alembic>=1.13.0
